requirements.txt
# Core dependencies
click>=8.0.0
rich>=13.0.0
pyyaml>=6.0
structlog>=23.0.0
requests>=2.28.0
aiohttp>=3.8.0

# AI and LangChain dependencies
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.1.0
langgraph>=0.1.0

# Security tools
python-nmap>=0.7.1

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0

# Environment management
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Networking utilities
netaddr>=0.8.0
dnspython>=2.3.0

# Async support
asyncio-mqtt>=0.13.0

# Logging and monitoring
loguru>=0.7.0

# Configuration validation
pydantic>=2.0.0

# CLI enhancements
typer>=0.9.0

# Progress bars and UI
tqdm>=4.65.0

# XML parsing (for Nmap output)
lxml>=4.9.0

# JSON handling
orjson>=3.8.0

# HTTP client enhancements
httpx>=0.24.0

# Cryptography utilities
cryptography>=41.0.0

# Time utilities
python-dateutil>=2.8.0

# File handling
pathlibutil>=0.2.0

# System utilities
psutil>=5.9.0